+++ packages/db/src/async_operations.ts
@@ -1,33 +1,19 @@
 import { toSnake } from 'snake-camel';
+import { ApiAsyncOperation } from '@cumulus/types/api/async_operations';
+import { PostgresAsyncOperation } from './types';
 
-interface CamelCaseAsyncOperation {
-  id: string
-  description: string
-  operationType: string
-  status: string
-  output?: object
-  taskArn?: string
-  createdAt?: Date
-  updatedAt?: Date
-}
-
-interface SnakeCaseAsyncOperation {
-  id: string
-  description: string
-  operation_type: string
-  status: string
-  output?: object
-  task_arn?: string
-  created_at?: Date
-  updated_at?: Date
-}
-
-export const translateAsyncOperationToSnakeCase = (
-  record: CamelCaseAsyncOperation
-): SnakeCaseAsyncOperation => {
-  const translatedRecord: SnakeCaseAsyncOperation = <SnakeCaseAsyncOperation>toSnake(record);
+export const translateApiAsyncOperationToPostgresAsyncOperation = (
+  record: ApiAsyncOperation
+): PostgresAsyncOperation => {
+  const translatedRecord: PostgresAsyncOperation = <PostgresAsyncOperation>toSnake(record);
   if (record.output !== undefined) {
     translatedRecord.output = record.output;
+  }
+  if (record.createdAt !== undefined) {
+    translatedRecord.created_at = new Date(record.createdAt);
+  }
+  if (record.updatedAt !== undefined) {
+    translatedRecord.updated_at = new Date(record.updatedAt);
   }
   return translatedRecord;
 };